requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
Flask-Session==0.5.0
Werkzeug==3.0.1
psycopg2-binary==2.9.9
requests==2.31.0
feedparser==6.0.10
beautifulsoup4==4.12.2
lxml==4.9.3
fuzzywuzzy==0.18.0
python-Levenshtein==0.21.1
